+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Added ability to exclude specific folders from duplicate scanning via settings page
 ### Fixed
 - Handle when there is no file to delete in bulk deletion
+- Fix an issue where the same file was displayed multiple times in duplicate
 
 ## 1.3.1 - 2024-12-26
 ### Added